requirements.txt
django
python-dotenv==1.0.1
djangorestframework==3.15.2
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
drf-spectacular==0.27.2
Pillow==10.0.0
django-celery-beat==2.6.0
cloudinary==1.39.1
python-dateutil==2.9.0
psycopg2-binary==2.9.9
djoser
django-filter
django-celery
django-celery-email
loguru==0.7.2
django-cors-headers
djangorestframework-simplejwt
requests
pycryptodome